# webapp/public/openapi.yaml
openapi: 3.0.3
info:
  title: Pakketpunten API
  version: 1.0.0
  description: |
    REST API for retrieving parcel point (pakketpunten) data for Dutch municipalities.

    This API provides access to comprehensive geospatial data about parcel pickup locations
    across the Netherlands, including:
    - **Parcel point locations** (coordinates, addresses, providers)
    - **Municipality boundaries**
    - **Provider statistics** (DHL, PostNL, DPD, VintedGo, Amazon, De Buren)

    Coverage areas (300/400/500 m) are not part of the municipality data: compute
    them from the points, or use the national unions at
    `/data/geo/coverage_300.geojson`, `coverage_400.geojson` and `coverage_500.geojson`.

    ## Features
    - ✅ No authentication required
    - ✅ CORS enabled for all origins
    - ✅ Rate limited (15 requests per hour)
    - ✅ GeoJSON format (RFC 7946)
    - ✅ Multiple identifier types (name, slug, CBS code)

    ## Data Format
    All responses are in GeoJSON format (FeatureCollection) with metadata.

    ## Rate Limiting
    - **Limit**: 15 requests per hour per IP address
    - **Headers**: `X-RateLimit-Limit`, `X-RateLimit-Remaining`
    - **Reset**: Rate limit resets every hour

    ## Data Updates
    Data is automatically updated weekly via GitHub Actions (Mondays at 02:00 UTC).

    ## Disclaimer
    **This API and all data are provided "as-is" without any warranties or guarantees.**

    The data is collected from various public sources and may contain inaccuracies, outdated information, or errors.
    We make no representations or warranties of any kind, express or implied, about:
    - The completeness, accuracy, reliability, or availability of the data
    - The suitability of the data for any particular purpose
    - The uptime or availability of the API

    Use this API and data at your own risk. The maintainers shall not be liable for any errors,
    omissions, or any losses, injuries, or damages arising from its use.

  contact:
    name: GitHub Repository
    url: https://github.com/robbertj85/pakketpunten
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://pakketpuntenviewer.nl/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server

tags:
  - name: Municipality Data
    description: Retrieve parcel point data for Dutch municipalities

paths:
  /municipalities:
    get:
      summary: List all municipalities
      description: |
        Returns the index of every municipality in the dataset: slug, CBS code,
        province, population and the current number of parcel points, split per
        carrier.

        Call this first — it gives you the valid identifiers for
        `/municipality/{identifier}` without having to guess slugs, and the
        counts let you pick the gemeenten you care about before downloading any
        GeoJSON.
      operationId: listMunicipalities
      tags:
        - Municipality Data
      responses:
        '200':
          description: The municipality index
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Number of municipalities returned
                    example: 343
                  municipalities:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: Zwolle
                        slug:
                          type: string
                          example: zwolle
                        province:
                          type: string
                          example: Overijssel
                        population:
                          type: integer
                          example: 130592
                        code:
                          type: string
                          nullable: true
                          description: CBS municipality code
                          example: GM0193
                        total_points:
                          type: integer
                          nullable: true
                          description: Null when the batch summary is unavailable
                          example: 214
                        carriers:
                          type: object
                          nullable: true
                          additionalProperties:
                            type: integer
                          description: Point count per carrier
                          example:
                            DHL: 62
                            PostNL: 58
                        url:
                          type: string
                          description: Path to the full data for this municipality
                          example: /api/v1/municipality/zwolle
        '429':
          description: Rate limit exceeded (60 requests per hour)

  /municipality/{identifier}:
    get:
      summary: Get municipality parcel point data
      description: |
        Retrieve complete parcel point data for a Dutch municipality.

        The `identifier` parameter accepts **multiple formats and variations**:

        **1. Municipality name** (case-insensitive):
        - Standard: `Amsterdam`, `Rotterdam`, `Utrecht`
        - Common names: `Den Haag`, `Den Bosch`
        - Official CBS names: `'s-Gravenhage`, `'s-Hertogenbosch`
        - English names: `The Hague`

        **2. URL slug**:
        - `amsterdam`, `den-haag`, `s-hertogenbosch`

        **3. CBS municipality code**:
        - `GM0363` (Amsterdam), `GM0518` (Den Haag), `GM0796` ('s-Hertogenbosch)

        **4. Alternative spellings** (supported via fuzzy matching):
        - With/without apostrophes: `s-Gravenhage`, `sgravenhage`
        - Regional identifiers: `Bergen Noord-Holland`, `Bergen (NH)`, `Bergen NH`
        - Abbreviated forms: `Hengelo (O)`, `Beek (L)`, `Rijswijk (ZH)`

        ### Supported Name Variations

        Common municipalities with multiple accepted names:
        - **Den Haag**: `'s-Gravenhage`, `s-Gravenhage`, `The Hague`, `Haag`
        - **'s-Hertogenbosch**: `Den Bosch`, `Bosch`, `shertogenbosch`
        - **Bergen (NH.)**: `Bergen Noord-Holland`, `Bergen (NH)`, `Bergen NH`
        - **Bergen (L.)**: `Bergen Limburg`, `Bergen (L)`, `Bergen L`
        - **Hengelo**: `Hengelo (O)`, `Hengelo Overijssel`
        - **Valkenburg aan de Geul**: `Valkenburg`, `Valkenburg (L)`

        Returns a GeoJSON FeatureCollection containing:
        - Parcel point locations (type: `pakketpunt`)
        - Municipality boundary (type: `boundary`)

      operationId: getMunicipalityData
      tags:
        - Municipality Data
      parameters:
        - name: identifier
          in: path
          required: true
          description: |
            Municipality identifier. Supports multiple formats:
            - Municipality name (standard or alternative spellings)
            - URL slug (e.g., `amsterdam`)
            - CBS code (e.g., `GM0363`)
          schema:
            type: string
          examples:
            by_name:
              summary: By municipality name
              value: Amsterdam
            by_common_name:
              summary: By common name (Den Haag)
              value: Den Haag
            by_official_name:
              summary: By official CBS name
              value: "'s-Gravenhage"
            by_english_name:
              summary: By English name
              value: The Hague
            by_alternative_name:
              summary: By alternative name (Den Bosch)
              value: Den Bosch
            by_regional_identifier:
              summary: By regional identifier
              value: Bergen Noord-Holland
            by_slug:
              summary: By URL slug
              value: amsterdam
            by_cbs_code:
              summary: By CBS code
              value: GM0363
            special_case_den_haag:
              summary: Den Haag (name)
              value: Den Haag
            special_case_s_hertogenbosch:
              summary: "'s-Hertogenbosch (name)"
              value: s-Hertogenbosch
      responses:
        "200":
          description: Successful response with GeoJSON data
          headers:
            X-RateLimit-Limit:
              description: Maximum number of requests allowed per time window
              schema:
                type: integer
                example: 100
            X-Municipality-Name:
              description: Official municipality name
              schema:
                type: string
                example: Amsterdam
            X-Municipality-Slug:
              description: URL-safe municipality slug
              schema:
                type: string
                example: amsterdam
            X-Municipality-Code:
              description: CBS municipality code
              schema:
                type: string
                example: GM0363
            Cache-Control:
              description: Cache control directive
              schema:
                type: string
                example: public, max-age=3600, stale-while-revalidate=86400
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/GeoJSONFeatureCollection"
        "400":
          description: Bad request - Invalid identifier or restricted municipality
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing_identifier:
                  summary: Missing identifier
                  value:
                    error: Missing identifier
                    message: Please provide a municipality identifier (name, slug, or CBS code)
                nederland_restricted:
                  summary: Nederland overview restricted
                  value:
                    error: Invalid municipality
                    message: National overview data is not available via the API. Please use individual municipality endpoints.
        "404":
          description: Municipality not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                not_found:
                  summary: Municipality not found
                  value:
                    error: Municipality not found
                    message: "No municipality found for identifier: InvalidName"
                    hint: "Try using a municipality name (e.g., Amsterdam), slug (e.g., amsterdam), or CBS code (e.g., GM0363)"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
                example: 900
            X-RateLimit-Limit:
              description: Maximum number of requests allowed per time window
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              description: Number of requests remaining in current time window
              schema:
                type: integer
                example: 0
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                rate_limit:
                  summary: Rate limit exceeded
                  value:
                    error: Rate limit exceeded
                    message: Maximum 100 requests per 15 minutes
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                internal_error:
                  summary: Server error
                  value:
                    error: Internal server error
                    message: An unexpected error occurred while processing your request

components:
  schemas:
    GeoJSONFeatureCollection:
      type: object
      required:
        - type
        - metadata
        - features
      properties:
        type:
          type: string
          enum: [FeatureCollection]
          description: GeoJSON type (always "FeatureCollection")
        metadata:
          type: object
          description: Municipality metadata
          properties:
            gemeente:
              type: string
              description: Municipality name
              example: Amsterdam
            slug:
              type: string
              description: URL-safe municipality slug
              example: amsterdam
            generated_at:
              type: string
              format: date-time
              description: Timestamp when data was generated
              example: "2025-01-15T10:30:00Z"
            total_points:
              type: integer
              description: Total number of parcel points
              example: 156
            providers:
              type: array
              items:
                type: string
                enum: [DHL, PostNL, VintedGo, DeBuren, DPD, Amazon]
              description: List of parcel providers in this municipality
              example: ["DHL", "PostNL", "DPD"]
            bounds:
              type: array
              items:
                type: number
              minItems: 4
              maxItems: 4
              description: Bounding box [minX, minY, maxX, maxY] in WGS84
              example: [4.72, 52.28, 5.07, 52.43]
        features:
          type: array
          description: GeoJSON features (parcel points, boundary)
          items:
            $ref: "#/components/schemas/GeoJSONFeature"

    GeoJSONFeature:
      type: object
      required:
        - type
        - properties
        - geometry
      properties:
        type:
          type: string
          enum: [Feature]
        properties:
          oneOf:
            - $ref: "#/components/schemas/PakketpuntProperties"
            - $ref: "#/components/schemas/BoundaryProperties"
        geometry:
          $ref: "#/components/schemas/GeoJSONGeometry"

    PakketpuntProperties:
      type: object
      description: Parcel point properties
      required:
        - type
        - locatieNaam
        - straatNaam
        - straatNr
        - vervoerder
        - latitude
        - longitude
      properties:
        type:
          type: string
          enum: [pakketpunt]
        locatieNaam:
          type: string
          description: Location name (e.g., store name)
          example: Albert Heijn Amsterdam Centraal
        straatNaam:
          type: string
          description: Street name
          example: Stationsplein
        straatNr:
          type: string
          description: Street number
          example: "15"
        vervoerder:
          type: string
          enum: [DHL, PostNL, VintedGo, DeBuren, DPD, Amazon]
          description: Parcel carrier/provider
          example: DHL
        puntType:
          type: string
          description: Type of parcel point
          example: ServicePoint
        bezettingsgraad:
          type: integer
          minimum: 0
          maximum: 100
          description: Occupancy rate (0-100%, randomly generated for demo)
          example: 45
        latitude:
          type: number
          format: double
          description: Latitude (WGS84)
          example: 52.379189
        longitude:
          type: number
          format: double
          description: Longitude (WGS84)
          example: 4.900272

    BoundaryProperties:
      type: object
      description: Municipality boundary properties
      required:
        - type
      properties:
        type:
          type: string
          enum: [boundary]
          description: Feature type (municipality boundary)

    GeoJSONGeometry:
      type: object
      description: GeoJSON geometry (Point, Polygon, or MultiPolygon)
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum: [Point, Polygon, MultiPolygon]
        coordinates:
          oneOf:
            - type: array
              items:
                type: number
              minItems: 2
              maxItems: 3
              description: Point coordinates [longitude, latitude]
            - type: array
              items:
                type: array
              description: Polygon or MultiPolygon coordinates

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: Municipality not found
        message:
          type: string
          description: Human-readable error message
          example: "No municipality found for identifier: InvalidName"
        hint:
          type: string
          description: Optional hint for resolving the error
          example: "Try using a municipality name (e.g., Amsterdam), slug (e.g., amsterdam), or CBS code (e.g., GM0363)"

externalDocs:
  description: GitHub Repository
  url: https://github.com/robbertj85/pakketpunten